refactor(context): Group context props and functions into a parent class
