+++ src/index.ts
@@ -3,7 +3,6 @@ import * as Util from './util';
 
 import API from './api';
 import SlashCommand from './command';
-import CommandContext from './context';
 import SlashCreator from './creator';
 import Server from './server';
 
@@ -24,6 +23,10 @@ import Message from './structures/message';
 import Permissions from './structures/permissions';
 import User from './structures/user';
 import UserFlags from './structures/userFlags';
+
+import ComponentContext from './structures/interfaces/componentContext';
+import CommandContext from './structures/interfaces/context';
+import MessageInteractionContext from './structures/interfaces/messageInteraction';
 
 const VERSION: string = require('../package').version;
 
@@ -70,8 +73,6 @@ export {
   ImageFormat
 } from './constants';
 
-export type { ConvertedOption } from './context';
-
 export { SlashCommandOptions, CommandPermissions } from './command';
 
 export {
@@ -90,11 +91,14 @@ export {
   EmbedVideo
 } from './structures/message';
 
+export type { ConvertedOption } from './structures/interfaces/context';
+
 export {
   API,
   BitField,
   Command,
   CommandContext,
+  ComponentContext,
   Constants,
   Context,
   Creator,
@@ -106,6 +110,7 @@ export {
   GCFServer,
   Member,
   Message,
+  MessageInteractionContext,
   Permissions,
   RequestHandler,
   SequentialBucket,
